requirements.txt
streamlit==1.30.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
ultralytics==8.0.196
Pillow==10.1.0 